+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   like this: `image1+image2+image3: [ package1, package2 ]`
 - Add `--verbose` flag to `list recipes` to display additional fields like version, arch, license and description.
 - Add `--verbose` flag to `list images`.
+- Add `--raw` flag to `list` subcommand to disable colored output.
 
 # 0.5.0
 - If at least one task fails to build **pkger** will now return with exit code 1 [#66](https://github.com/vv9k/pkger/pull/66)